requirements.txt
fastapi[all]
SQLAlchemy
mysql-connector-python
python-dotenv
azure-core
azure-storage-blob
pandas
pymysql
flask